+++ packages/components/src/__tests__/RadioGroup.spec.tsx
@@ -2,7 +2,6 @@ import 'jest';
 import * as React from 'react';
 import { shallow } from 'enzyme';
 import RadioGroup from '../components/RadioGroup/RadioGroup';
-import RadioOption from '../components/RadioOption/RadioOption';
 import StyledRadioGroup from '../components/RadioGroup/RadioGroup.style';
 import { darkTheme } from '../theme';
 
@@ -14,28 +13,13 @@ describe('RadioGroup', () => {
         name="themes"
         value={name}
         onChange={val => console.log(val)}
-      >
-        <RadioOption label="default" value="default" />
-        <RadioOption label="primary" value="primary" />
-      </RadioGroup>
+        options={[
+          { label: 'default', value: 'default' },
+          { label: 'primary', value: 'primary' }
+        ]}
+      />
     );
 
-    expect(component).toMatchSnapshot();
-  });
-
-  it('RadioGroup should render nothing', () => {
-    const component = shallow(
-      <RadioGroup
-        label="Active Theme"
-        name="themes"
-        value={name}
-        onChange={val => console.log(val)}
-      >
-        <p>Another child</p>
-      </RadioGroup>
-    );
-
-    expect(component).toMatchObject({});
     expect(component).toMatchSnapshot();
   });