+++ packages/components/src/components/Counter/Counter.tsx
@@ -5,7 +5,7 @@ import StyledCounter from './Counter.style';
 import Label from '../Label/Label';
 
 export interface Props {
-  iconBefore?: JSX.Element;
+  iconBefore?: HTMLElement;
   label: string;
   onChange: (val: number, suffix: string | undefined) => void;
   value: number;
@@ -18,7 +18,7 @@ export interface Props {
 
 const countDecimals = (number: number) => {
   if (Math.floor(number) === number) return 0;
-  return number.toString().split('.')[1].length;
+  return number.toString().split('.')[1].length || 0;
 };
 
 const Counter: React.FC<Props> = ({