+++ packages/components/src/components/Panel/Panel.tsx
@@ -5,7 +5,6 @@ import { ThemePlaygroundContext } from '../../contexts/ThemePlaygroundProvider';
 import Code from '../Code/Code';
 import SettingsItem from '../SettingsItem/SettingsItem';
 import RadioGroup from '../RadioGroup/RadioGroup';
-import RadioOption from '../RadioOption/RadioOption';
 
 import StyledPanel from './Panel.style';
 
@@ -31,11 +30,8 @@ const Panel = ({ className }: { className?: string | null }) => {
                   onChange={val =>
                     updateActiveTheme(themes.filter(t => t.name === val)[0])
                   }
-                >
-                  {themes.map(t => (
-                    <RadioOption key={t.name} label={t.name} value={t.name} />
-                  ))}
-                </RadioGroup>
+                  options={themes.map(t => ({ label: t.name, value: t.name }))}
+                />
               </div>
             )}
             <SettingsItem />