+++ packages/components/src/components/RadioGroup/RadioGroup.tsx
@@ -1,8 +1,14 @@
 import * as React from 'react';
 
 import StyledRadioGroup from './RadioGroup.style';
-import RadioOption, { RadioOptionProps } from '../RadioOption/RadioOption';
+import RadioOption from '../RadioOption/RadioOption';
 import Label from '../Label/Label';
+
+type RadioOptions = {
+  label: string;
+  value: string;
+};
+
 export interface RadioGroupProps {
   iconBefore?: JSX.Element;
   label: string;
@@ -10,7 +16,7 @@ export interface RadioGroupProps {
   onChange: (val: string) => void;
   value: string;
   name: string;
-  children?: React.ReactNode;
+  options: RadioOptions[];
 }
 
 const RadioGroup: React.FC<RadioGroupProps> = ({
@@ -20,26 +26,23 @@ const RadioGroup: React.FC<RadioGroupProps> = ({
   name,
   value,
   onChange,
-  children
+  options
 }) => {
   return (
     <StyledRadioGroup>
       <Label iconBefore={iconBefore} label={label} description={description} />
 
-      {React.Children.map(
-        children,
-        (child: React.ReactElement<RadioOptionProps>) => {
-          if (child.type !== RadioOption) return;
-          return React.cloneElement(child, {
-            onChange,
-            name,
-            label: child.props.label,
-            value: child.props.value,
-            icon: child.props.icon,
-            isChecked: child.props.value === value
-          });
-        }
-      )}
+      {options &&
+        options.map(item => (
+          <RadioOption
+            key={item.value}
+            label={item.label}
+            value={item.value}
+            name={name}
+            onChange={() => onChange(item.value)}
+            isChecked={value === item.value}
+          />
+        ))}
     </StyledRadioGroup>
   );
 };