+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `Select` and `RadioGroup` to override components
+
+### Fixed
+
+- `Counter` value handling
+
 ## [0.1.3] - 2020-02-20
 
 ### Fixed